Move ResolverError over to exceptions where it seems to be long.
