+++ tests/test_resolver.py
@@ -1,6 +1,7 @@
 import connexion.app
 from connexion.operation import Operation
-from connexion.resolver import Resolver, RestyResolver, ResolverError
+from connexion.exceptions import ResolverError
+from connexion.resolver import Resolver, RestyResolver
 
 import pytest
